Performs optimal binning of a numeric variable for Weight of Evidence (WoE) and Information Value (IV) using a decision tree-based approach

This function processes a numeric variable and applies a decision tree algorithm to iteratively split the data into bins. The tree splits are based on Information Value (IV) gain, with constraints on minimum bin size and maximum depth. The function ensures monotonicity in event rates, merging bins if necessary to reduce the number of bins to the specified maximum. It also calculates WoE and IV for the generated bins.

Usage

OptimalBinningNumericTree(
  target,
  feature,
  max_bins = 7L,
  lambda = 0.1,
  min_bin_size = 0.05,
  min_iv_gain = 0.01,
  max_depth = 10L,
  monotonicity_direction = "increase"
)

Arguments

target

Integer vector representing the binary target variable, where 1 indicates a positive event (e.g., default) and 0 indicates a negative event (e.g., non-default).

feature

Numeric vector representing the numeric variable to be binned.

max_bins

(Optional) Maximum number of bins to generate. Default is 7.

lambda

(Optional) Regularization parameter to penalize tree splits. Default is 0.1.

min_bin_size

(Optional) Minimum size a bin must have as a proportion of the total data. Default is 0.05.

min_iv_gain

(Optional) Minimum Information Value (IV) gain required to perform a split. Default is 0.01.

max_depth

(Optional) Maximum depth of the decision tree. Default is 10.

monotonicity_direction

(Optional) String that defines the monotonicity direction of event rates, either "increase" for increasing monotonicity or "decrease" for decreasing monotonicity. Default is "increase".

Value

A list with the following elements:

feature_woe: Numeric vector with the WoE assigned to each instance of the processed numeric variable.

bin: DataFrame with the generated bins, containing the following fields:

bin: String representing the range of values for each bin.

woe: Weight of Evidence (WoE) for each bin.

count: Total number of observations in each bin.

count_pos: Count of positive events in each bin.

count_neg: Count of negative events in each bin.

woe: Numeric vector with the WoE for each bin.

iv: Total Information Value (IV) calculated for the variable.

pos: Vector with the count of positive events in each bin.

neg: Vector with the count of negative events in each bin.