Optimal Binning and Weight of Evidence Calculation

This function performs optimal binning and calculates Weight of Evidence (WoE) for both numerical and categorical features.

Usage

obwoe(
  dt,
  target,
  features = NULL,
  method = "auto",
  preprocess = TRUE,
  min_bins = 3,
  max_bins = 4,
  control = list(),
  positive = "bad|1"
)

Arguments

dt

A data.table containing the dataset.

target

The name of the target variable (must be binary).

features

Vector of feature names to process. If NULL, all features except the target will be processed.

method

The binning method to use. Can be "auto" or a specific method name.

preprocess

Logical. Whether to preprocess the data before binning.

min_bins

Minimum number of bins.

max_bins

Maximum number of bins.

control

A list of additional control parameters.

positive

Character string specifying which category should be considered as positive. Must be either "bad|1" or "good|1".

Value

A list containing:

woedt

The original dataset with added WoE columns

woebins

Information about the bins created

prepreport

Preprocessing report for each feature

bestsreport

Report on the best models used

failedfeatures

List of features that failed processing

bestmethod

Best method used for binning