Stepwise Model Selection for oblr Objects

Performs stepwise model selection (backward, forward, both) for oblr model objects based on the Akaike Information Criterion (AIC).

Usage

# S3 method for class 'oblr'
step(
  object,
  scope = NULL,
  scale = 0,
  direction = c("both", "backward", "forward"),
  trace = 1,
  keep = NULL,
  steps = 1000,
  k = 2,
  ...
)

Arguments

object

An object of class oblr.

scope

A two-sided formula defining the range of models examined in the stepwise search. Can be a single formula or a list with lower and upper components. If missing, the scope is determined based on the current model.

scale

Unused. Set to 0 by default.

direction

The direction of stepwise search. Must be one of "both", "backward", or "forward".

trace

An integer indicating the level of tracing information to be printed. 0 for no tracing, 1 for minimal tracing, and higher values for more detailed tracing.

keep

An optional function to be applied to each model as it is evaluated. Useful for tracking model selection history. Set to NULL by default.

steps

The maximum number of steps to perform. Default is 1000.

k

The penalty per parameter to be used in the AIC calculation. Default is 2.

...

Additional arguments passed to or from other methods.

Value

An oblr object representing the final selected model after stepwise selection. Attributes include the selection path and AIC history if keep is utilized.

Details

The step.oblr function iteratively adds or removes predictors from an oblr model to optimize the AIC. It supports three directions of selection:

Backward: Starts with the full model and removes predictors.

Forward: Starts with a minimal model and adds predictors.

Both: Combines both backward elimination and forward selection.

The function evaluates all possible single-step additions or removals of predictors and selects the model that provides the best (lowest) AIC at each step. The process continues until no further improvement in AIC is possible or the maximum number of steps is reached.

Examples

if (FALSE) { # \dontrun{
# Assuming 'dt' is your data frame and 'oblr' is your model fitting function
library(Matrix)

# Fit the initial full model
m_full <- oblr(y ~ X1 + X2 + X3, data = dt)

# Perform backward stepwise selection
step_backward <- step(m_full, direction = "backward")

# Perform forward stepwise selection with a minimal model
m_min <- oblr(y ~ 1, data = dt)
step_forward <- step(m_min, scope = list(lower = ~1, upper = ~ X1 + X2 + X3), direction = "forward")

# Perform both directions stepwise selection
step_both <- step(m_full, direction = "both")
} # }